posco-dx-agent-dev-day5/pyproject.toml
[project]
name = "base-agent"
version = "0.1.0"
description = "나만의 AI 에이전트"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.12",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "fastmcp>=3.1.1",
    "chainlit>=2.5.0",
    "python-dotenv>=1.1.0",
    "numpy>=2.0.0",
    "pypdf>=4.0.0",
    "langfuse>=4.0.1",
]
